+++ cli/internal/metadataobject/actions/cli_extension/cli_extension.go
@@ -12,7 +12,7 @@ import (
 	"github.com/hasura/graphql-engine/cli/v2/internal/metadataobject/actions/types"
 	"github.com/pkg/errors"
 	"github.com/sirupsen/logrus"
-	"gopkg.in/yaml.v2"
+	"gopkg.in/yaml.v3"
 )
 
 // Config represents the object to interact with cli-ext