+++ cli/internal/metadataobject/graphql_schema_introspection/graphql_schema_introspection.go
@@ -1,6 +1,7 @@
 package graphqlschemaintrospection
 
 import (
+	"bytes"
 	"io/ioutil"
 	"path/filepath"
 
@@ -9,7 +10,7 @@ import (
 	"github.com/sirupsen/logrus"
 
 	"github.com/hasura/graphql-engine/cli/v2"
-	"gopkg.in/yaml.v2"
+	"gopkg.in/yaml.v3"
 )
 
 type MetadataObject struct {
@@ -42,53 +43,65 @@ func (o *MetadataObject) CreateFiles() error {
 	return nil
 }
 
-func (o *MetadataObject) Build(metadata *yaml.MapSlice) metadataobject.ErrParsingMetadataObject {
-	data, err := ioutil.ReadFile(filepath.Join(o.MetadataDir, o.Filename()))
-	if err != nil {
-		return o.error(err)
-	}
-	item := yaml.MapItem{
-		Key: o.Key(),
-	}
-	var obj yaml.MapSlice
-	err = yaml.Unmarshal(data, &obj)
-	if err != nil {
-		return o.error(err)
-	}
-	if len(obj) > 0 {
-		item.Value = obj
-		*metadata = append(*metadata, item)
-	}
-	return nil
+type graphQLSchemaIntrospectionObject struct {
+	DisabledForRoles []yaml.Node `yaml:"disabled_for_roles"`
 }
 
-func (o *MetadataObject) Export(metadata yaml.MapSlice) (map[string][]byte, metadataobject.ErrParsingMetadataObject) {
-	var graphqlSchemaIntroSpection interface{}
-	for _, item := range metadata {
-		k, ok := item.Key.(string)
-		if !ok || k != o.Key() {
-			continue
+func (o *MetadataObject) Build() (map[string]interface{}, metadataobject.ErrParsingMetadataObject) {
+	data, err := ioutil.ReadFile(filepath.Join(o.MetadataDir, o.Filename()))
+	if err != nil {
+		return nil, o.error(err)
+	}
+	var obj graphQLSchemaIntrospectionObject
+	err = yaml.Unmarshal(data, &obj)
+	if err != nil {
+		return nil, o.error(err)
+	}
+
+	return map[string]interface{}{o.Key(): obj}, nil
+}
+
+// Export implementation here can probably use metadataobject.DefaultExport
+// the only thing stopping us from using it is the metadata API behaviour for this object.
+// For other objects an empty value ([] or {}) depending on what type of object will
+// "reset" it. ie If I wanted to clear all the cron triggers  on the server I could send metadata
+// JSON with "cron_triggers": []. Similarly, If I try to "clear" "graphql_schema_introspection" it'll not work
+//
+// $ hasura metadata apply -o json --dry-run
+// {
+//  // ....other objects skipped for brevity
+//  "graphql_schema_introspection": {}
+//}
+// $ hasura md apply
+// FATA[0001] error applying metadata
+//{
+//  "path": "$.args.metadata.graphql_schema_introspection",
+//  "error": "the key 'disabled_for_roles' was not present",
+//  "code": "parse-failed"
+//}
+func (o *MetadataObject) Export(metadata map[string]yaml.Node) (map[string][]byte, metadataobject.ErrParsingMetadataObject) {
+	var object graphQLSchemaIntrospectionObject
+	if v, ok := metadata[o.Key()]; ok {
+		objectbs, err := yaml.Marshal(v)
+		if err != nil {
+			return nil, o.error(err)
 		}
-		graphqlSchemaIntroSpection = item.Value
+		if err := yaml.Unmarshal(objectbs, &object); err != nil {
+			return nil, o.error(err)
+		}
 	}
-	if graphqlSchemaIntroSpection == nil {
-		o.logger.WithFields(logrus.Fields{
-			"object": o.Key(),
-			"reason": "not found in metadata",
-		}).Debugf("skipped building %s", o.Key())
-		return nil, nil
-	}
-	data, err := yaml.Marshal(graphqlSchemaIntroSpection)
+	var buf bytes.Buffer
+	err := metadataobject.GetEncoder(&buf).Encode(object)
 	if err != nil {
 		return nil, o.error(err)
 	}
 	return map[string][]byte{
-		filepath.Join(o.MetadataDir, o.Filename()): data,
+		filepath.ToSlash(filepath.Join(o.BaseDirectory(), o.Filename())): buf.Bytes(),
 	}, nil
 }
 
 func (o *MetadataObject) Key() string {
-	return "graphql_schema_introspection"
+	return metadataobject.GraphQLSchemaIntrospectionKey
 }
 
 func (o *MetadataObject) Filename() string {