+++ cli/internal/metadataobject/inherited_roles/inherited_roles.go
@@ -9,7 +9,7 @@ import (
 	"github.com/sirupsen/logrus"
 
 	"github.com/hasura/graphql-engine/cli/v2"
-	"gopkg.in/yaml.v2"
+	"gopkg.in/yaml.v3"
 )
 
 type InheritedRolesConfig struct {
@@ -35,49 +35,25 @@ func (ir *InheritedRolesConfig) CreateFiles() error {
 	return nil
 }
 
-func (ir *InheritedRolesConfig) Build(metadata *yaml.MapSlice) metadataobject.ErrParsingMetadataObject {
+func (ir *InheritedRolesConfig) Build() (map[string]interface{}, metadataobject.ErrParsingMetadataObject) {
 	data, err := ioutil.ReadFile(filepath.Join(ir.MetadataDir, ir.Filename()))
-	if err != nil {
-		return ir.error(err)
-	}
-	item := yaml.MapItem{
-		Key:   "inherited_roles",
-		Value: []yaml.MapSlice{},
-	}
-	err = yaml.Unmarshal(data, &item.Value)
-	if err != nil {
-		return ir.error(err)
-	}
-	*metadata = append(*metadata, item)
-	return nil
-}
-
-func (ir *InheritedRolesConfig) Export(metadata yaml.MapSlice) (map[string][]byte, metadataobject.ErrParsingMetadataObject) {
-	var inheritedRoles interface{}
-	for _, item := range metadata {
-		k, ok := item.Key.(string)
-		if !ok || k != "inherited_roles" {
-			continue
-		}
-		inheritedRoles = item.Value
-	}
-	if inheritedRoles == nil {
-		ir.logger.WithFields(logrus.Fields{
-			"reason": "not found in metadata",
-		}).Debugf("skipped building %s", ir.Key())
-		return nil, nil
-	}
-	data, err := yaml.Marshal(inheritedRoles)
 	if err != nil {
 		return nil, ir.error(err)
 	}
-	return map[string][]byte{
-		filepath.ToSlash(filepath.Join(ir.MetadataDir, ir.Filename())): data,
-	}, nil
+	var obj []yaml.Node
+	err = yaml.Unmarshal(data, &obj)
+	if err != nil {
+		return nil, ir.error(err)
+	}
+	return map[string]interface{}{ir.Key(): obj}, nil
+}
+
+func (ir *InheritedRolesConfig) Export(metadata map[string]yaml.Node) (map[string][]byte, metadataobject.ErrParsingMetadataObject) {
+	return metadataobject.DefaultExport(ir, metadata, ir.error, metadataobject.DefaultObjectTypeSequence)
 }
 
 func (ir *InheritedRolesConfig) Key() string {
-	return "inherited_roles"
+	return metadataobject.InheritedRolesKey
 }
 
 func (ir *InheritedRolesConfig) Filename() string {