+++ cli/internal/projectmetadata/handler.go
@@ -6,20 +6,23 @@ import (
 	"fmt"
 	"io"
 	"io/fs"
+	"io/ioutil"
 	"os"
 	"path/filepath"
 
+	goyaml "github.com/goccy/go-yaml"
+
 	"github.com/hasura/graphql-engine/cli/v2/internal/metadataobject"
+	"github.com/hasura/graphql-engine/cli/v2/internal/metadatautil"
 
 	"github.com/hasura/graphql-engine/cli/v2"
 
-	gyaml "github.com/goccy/go-yaml"
 	"github.com/hasura/graphql-engine/cli/v2/internal/hasura"
 	"github.com/mitchellh/mapstructure"
 	"github.com/pkg/errors"
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/afero"
-	"gopkg.in/yaml.v2"
+	"gopkg.in/yaml.v3"
 )
 
 // Handler will be responsible for interaction between a hasura instance and Objects
@@ -67,8 +70,20 @@ func (h *Handler) ExportMetadata() (map[string][]byte, error) {
 	if err != nil {
 		return nil, err
 	}
-	var c yaml.MapSlice
-	err = yaml.NewDecoder(resp).Decode(&c)
+	jsonmdbs, err := ioutil.ReadAll(resp)
+	if err != nil {
+		return nil, err
+	}
+	// We don't want to strongly type metadata here, but we cannot use a catch-all kind of datastructures like
+	// map[string]interface{} here because it'll mess up the ordering.
+	// So we directly translate JSON to YAML, to preserve it.
+	yamlmdbs, err := metadatautil.JSONToYAML(jsonmdbs)
+	if err != nil {
+		return nil, err
+	}
+
+	var c map[string]yaml.Node
+	err = yaml.NewDecoder(bytes.NewReader(yamlmdbs)).Decode(&c)
 	if err != nil {
 		return nil, err
 	}
@@ -97,39 +112,56 @@ func (h *Handler) ReloadMetadata() (io.Reader, error) {
 	return r, err
 }
 
-func (h *Handler) BuildMetadata() (yaml.MapSlice, error) {
-	var tmpMeta yaml.MapSlice
+func (h *Handler) buildMetadataMap() (map[string]interface{}, error) {
+	var metadata = map[string]interface{}{}
 	for _, object := range h.objects {
-		err := object.Build(&tmpMeta)
+		objectMetadata, err := object.Build()
 		if err != nil {
 			if errors.Is(err, fs.ErrNotExist) {
 				h.logger.Debugf("metadata file for %s was not found, assuming an empty file", object.Key())
 				continue
 			}
-			return tmpMeta, errors.Wrap(err, fmt.Sprintf("cannot build %s from project", object.Key()))
+			return nil, fmt.Errorf("cannot build %s from project: %w", object.Key(), err)
+		}
+		for k, v := range objectMetadata {
+			metadata[k] = v
 		}
 	}
-	return tmpMeta, nil
+	return metadata, nil
 }
 
-func (h *Handler) MakeJSONMetadata() ([]byte, error) {
-	tmpMeta, err := h.BuildMetadata()
+// buildMetadata is a private function because we don't intend consumers of this package
+// to use the returned result (metadataobject.Metadata) directly because they may assume that they can use
+// json.Marshal to get JSON representation of the built metadata. But this assumption will not hold true because
+// the underlying types might have instances of yaml.Node which is not friendly with a json.Marshal and can produce
+// unexpected results. Rather to get a JSON / YAML form of built metadata make use of Handler.BuildYAMLMetadata and
+// Handler.BuildJSONMetadata helper functions
+func (h *Handler) buildMetadata() (*Metadata, error) {
+	metadataMap, err := h.buildMetadataMap()
 	if err != nil {
 		return nil, err
 	}
-	yByt, err := yaml.Marshal(tmpMeta)
+	return GenMetadataFromMap(metadataMap)
+}
+
+func (h *Handler) BuildYAMLMetadata() ([]byte, error) {
+	metadata, err := h.buildMetadata()
 	if err != nil {
 		return nil, err
 	}
-	jbyt, err := gyaml.YAMLToJSON(yByt)
+	return metadata.YAML()
+}
+
+func (h *Handler) BuildJSONMetadata() ([]byte, error) {
+	metadata, err := h.buildMetadata()
 	if err != nil {
 		return nil, err
 	}
-	return jbyt, nil
+	return metadata.JSON()
 }
 
 func (h *Handler) V1ApplyMetadata() (io.Reader, error) {
-	jbyt, err := h.MakeJSONMetadata()
+	jbyt, err := h.BuildJSONMetadata()
 	if err != nil {
 		return nil, err
 	}
@@ -141,7 +173,7 @@ func (h *Handler) V1ApplyMetadata() (io.Reader, error) {
 }
 
 func (h *Handler) V2ApplyMetadata() (*hasura.V2ReplaceMetadataResponse, error) {
-	jbyt, err := h.MakeJSONMetadata()
+	jbyt, err := h.BuildJSONMetadata()
 	if err != nil {
 		return nil, err
 	}
@@ -252,4 +284,61 @@ func (obj InconsistentMetadataObject) GetReason() string {
 		return fmt.Sprintf("%.80s...", string(b))
 	}
 	return "N/A"
+}
+
+// Metadata does not strictly mirror the actual structure of server metadata
+// this is evident in the struct below, because V3 metadata does not contain "tables" / "functions" key
+//
+// this is rather a utility / helper struct which allow us to unmarshal / marshal
+// metadata bytes in a specific order.
+type Metadata struct {
+	Version          interface{} `yaml:"version" mapstructure:"version"`
+	Sources          interface{} `yaml:"sources,omitempty" mapstructure:"sources,omitempty"`
+	Tables           interface{} `yaml:"tables,omitempty" mapstructure:"tables,omitempty"`
+	Functions        interface{} `yaml:"functions,omitempty" mapstructure:"functions,omitempty"`
+	Actions          interface{} `yaml:"actions,omitempty" mapstructure:"actions,omitempty"`
+	CustomTypes      interface{} `yaml:"custom_types,omitempty" mapstructure:"custom_types,omitempty"`
+	RemoteSchemas    interface{} `yaml:"remote_schemas,omitempty" mapstructure:"remote_schemas,omitempty"`
+	QueryCollections interface{} `yaml:"query_collections,omitempty" mapstructure:"query_collections,omitempty"`
+	AllowList        interface{} `yaml:"allowlist,omitempty" mapstructure:"allowlist,omitempty"`
+	CronTriggers     interface{} `yaml:"cron_triggers,omitempty" mapstructure:"cron_triggers,omitempty"`
+	Network          interface{} `yaml:"network,omitempty" mapstructure:"network,omitempty"`
+	APILimits        interface{} `yaml:"api_limits,omitempty" mapstructure:"api_limits,omitempty"`
+	RestEndpoints    interface{} `yaml:"rest_endpoints,omitempty" mapstructure:"rest_endpoints,omitempty"`
+	InheritedRoles   interface{} `yaml:"inherited_roles,omitempty" mapstructure:"inherited_roles,omitempty"`
+
+	// HGE Pro
+	GraphQLSchemaIntrospection interface{} `yaml:"graphql_schema_introspection,omitempty" mapstructure:"graphql_schema_introspection,omitempty"`
+}
+
+// JSON is a helper function which returns JSON representation of Metadata
+// This exists because we cannot directly do a json.Marshal on Metadata because
+// the underlying type of struct fields might be yaml.Node
+// which might give unintended result
+func (m Metadata) JSON() ([]byte, error) {
+	yamlbs, err := m.YAML()
+	if err != nil {
+		return nil, err
+	}
+	jsonbs, err := goyaml.YAMLToJSON(yamlbs)
+	if err != nil {
+		return nil, err
+	}
+	return jsonbs, nil
+}
+
+func (m Metadata) YAML() ([]byte, error) {
+	var buf bytes.Buffer
+	if err := metadataobject.GetEncoder(&buf).Encode(m); err != nil {
+		return nil, err
+	}
+	return buf.Bytes(), nil
+}
+
+func GenMetadataFromMap(metadata map[string]interface{}) (*Metadata, error) {
+	var m = new(Metadata)
+	if err := mapstructure.Decode(metadata, m); err != nil {
+		return nil, err
+	}
+	return m, nil
 }