+++ cli/internal/projectmetadata/handler_test.go
@@ -1,7 +1,19 @@
 package projectmetadata
 
 import (
+	"bytes"
 	"encoding/json"
+	"fmt"
+	"github.com/hasura/graphql-engine/cli/v2"
+	"github.com/hasura/graphql-engine/cli/v2/internal/cliext"
+	"github.com/hasura/graphql-engine/cli/v2/internal/metadataobject"
+	"github.com/hasura/graphql-engine/cli/v2/version"
+	"github.com/mitchellh/go-homedir"
+	"github.com/sirupsen/logrus"
+	"github.com/stretchr/testify/assert"
+	"io/ioutil"
+	"os"
+	"path/filepath"
 	"testing"
 )
 
@@ -343,6 +355,92 @@ func Test_inconsistentObject_GetReason(t *testing.T) {
 			if got := obj.GetReason(); got != tt.want {
 				t.Errorf("GetReason() = %v, want %v", got, tt.want)
 			}
+		})
+	}
+}
+
+func TestGenMetadataFromMap(t *testing.T) {
+	testEC := cli.NewExecutionContext()
+	testEC.Version = version.NewCLIVersion(version.DevVersion)
+	testEC.Logger = logrus.New()
+	home, err := homedir.Dir()
+	assert.NoError(t, err)
+	testEC.GlobalConfigDir = filepath.Join(home, cli.GlobalConfigDirName)
+	testEC.Config = &cli.Config{Version: cli.V3}
+	testEC.HasMetadataV3 = true
+	assert.NoError(t, cliext.Setup(testEC))
+
+	getObjects := func(metadataDir string) metadataobject.Objects {
+		return GetMetadataObjectsWithDir(testEC, metadataDir)
+	}
+	type args struct {
+		metadata map[string]interface{}
+	}
+	tests := []struct {
+		id         string
+		name       string
+		args       args
+		wantGolden string
+		wantErr    assert.ErrorAssertionFunc
+	}{
+		{
+			"t1",
+			"can generate metadata struct for metadata v3",
+			args{
+				metadata: func() map[string]interface{} {
+					metadataDir := filepath.Join("testdata/test-gen-metadata-from-map/t1/metadata")
+					testEC.MetadataDir = metadataDir
+					handler := NewHandler(getObjects(metadataDir), nil, nil, logrus.New())
+					assert.NoError(t, err)
+					metadataMap, err := handler.buildMetadataMap()
+					assert.NoError(t, err)
+					return metadataMap
+				}(),
+			},
+			"",
+			assert.NoError,
+		},
+		{
+			"t2",
+			"can generate metadata struct for metadata v2",
+			args{
+				metadata: func() map[string]interface{} {
+					// settings for config v2
+					testEC.Config.Version = cli.V2
+					testEC.HasMetadataV3 = false
+
+					metadataDir := filepath.Join("testdata/test-gen-metadata-from-map/t2/metadata")
+					testEC.MetadataDir = metadataDir
+					handler := NewHandler(getObjects(metadataDir), nil, nil, logrus.New())
+					assert.NoError(t, err)
+					metadataMap, err := handler.buildMetadataMap()
+					assert.NoError(t, err)
+					return metadataMap
+				}(),
+			},
+			"",
+			assert.NoError,
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			got, err := GenMetadataFromMap(tt.args.metadata)
+			if !tt.wantErr(t, err, fmt.Sprintf("GenMetadataFromMap(%v)", tt.args.metadata)) {
+				return
+			}
+			gotJSON, err := got.JSON()
+			var buf bytes.Buffer
+			assert.NoError(t, json.Indent(&buf, gotJSON, "", "  "))
+			assert.NoError(t, err)
+			goldenFile := filepath.Join("testdata/test-gen-metadata-from-map", tt.id, "want.json")
+
+			// uncomment to update golden file
+			assert.NoError(t, ioutil.WriteFile(goldenFile, buf.Bytes(), os.ModePerm))
+
+			wantbs, err := ioutil.ReadFile(goldenFile)
+			assert.NoError(t, err)
+
+			assert.Equalf(t, string(wantbs), buf.String(), "GenMetadataFromMap(%v)", tt.args.metadata)
 		})
 	}
 }